+++ src/views/home/creation/components/NodeEditor.vue
@@ -1,5 +1,7 @@
 <template>
   <div class="node-edit-box">
+    <el-button type="success" icon="el-icon-upload" round @click="updateNodeXY" style="margin-bottom: 22px">保存结点布局
+    </el-button>
     <div style="font-size: 24px;font-weight: 900;margin-bottom: 20px">结点编辑</div>
     <div>
       <img class="img-box" :src="currentNode.data.url"/>
@@ -9,7 +11,7 @@
           <el-descriptions-item label="结点信息">{{ currentNode.data.content }}</el-descriptions-item>
         </el-descriptions>
       </div>
-  </div>
+    </div>
     <div>
       <div class="node-edit-color">
         字体颜色 <i class="el-icon-s-flag" :style="{'color':currentNode.fontColor}"/>
@@ -22,8 +24,9 @@
       </div>
     </div>
     <el-row>
-      <el-tooltip class="item" effect="dark" content="删除结点" placement="bottom">
-        <el-button type="danger" icon="el-icon-delete" circle v-if="currentNode.data.ableDelete" @click="deleteNode(currentNode)"></el-button>
+      <el-tooltip class="item" effect="dark" content="删除结点" placement="bottom" v-if="currentNode.data.ableDelete">
+        <el-button type="danger" icon="el-icon-delete" circle
+                   @click="deleteNode(currentNode)"></el-button>
       </el-tooltip>
       <el-tooltip class="item" effect="dark" content="编辑结点" placement="bottom">
         <el-button type="primary" icon="el-icon-edit" circle @click="changeNode(currentNode)"></el-button>
@@ -31,24 +34,26 @@
       <el-tooltip class="item" effect="dark" content="添加结点" placement="bottom">
         <el-button type="warning" icon="el-icon-plus" circle @click="newNodeDialogVisible=true"></el-button>
       </el-tooltip>
-      <el-tooltip class="item" effect="dark" content="保存布局" placement="bottom">
-        <el-button type="success" icon="el-icon-check" circle @click="updateNodeXY"></el-button>
-      </el-tooltip>
+
     </el-row>
 
     <el-dialog title="修改结点信息" :visible.sync="dialogVisible" width="30%" :close-on-click-modal=false>
-      <el-form :model="curNode" :rules="rules" :ref="curNode">
+      <el-form :model="curNode" :rules="rulesForUpdate" :ref="curNode">
         <img class="img-box" :src="curNode.url" v-show="msgFormSon" style="margin-bottom: 10px;float: inherit;"/>
-        <cropper-image @imgUploaded="updateNodePic($event, curNode)" @func="getMsgFormSon" style="margin-bottom: 10px;"></cropper-image>
+        <cropper-image @imgUploaded="updateNodePic($event, curNode)" @func="getMsgFormSon"
+                       style="margin-bottom: 10px;"></cropper-image>
         <div style="height: 80px">
           <div class="node-edit-color" style="margin: 22px;">
-            字体颜色 <el-color-picker v-model="curNode.fontColor" show-alpha :predefine="predefineColors"/>
+            字体颜色
+            <el-color-picker v-model="curNode.fontColor" show-alpha :predefine="predefineColors"/>
           </div>
           <div class="node-edit-color" style="margin: 22px;">
-            背景颜色 <el-color-picker v-model="curNode.color" show-alpha :predefine="predefineColors"/>
+            背景颜色
+            <el-color-picker v-model="curNode.color" show-alpha :predefine="predefineColors"/>
           </div>
           <div class="node-edit-color" style="margin: 22px;">
-            边框颜色 <el-color-picker v-model="curNode.borderColor" show-alpha :predefine="predefineColors"/>
+            边框颜色
+            <el-color-picker v-model="curNode.borderColor" show-alpha :predefine="predefineColors"/>
           </div>
         </div>
         <el-form-item prop="id">
@@ -63,18 +68,22 @@
     </el-dialog>
 
     <el-dialog title="添加新结点" :visible.sync="newNodeDialogVisible" width="30%" :close-on-click-modal=false>
-      <el-form :model="newNodeForm" :rules="rules" ref="newNodeForm">
+      <el-form :model="newNodeForm" :rules="rulesForAdd" ref="newNodeForm">
         <img class="img-box" :src="newNodeForm.url" v-show="msgFormSon" style="margin-bottom: 10px;float: inherit;"/>
-        <cropper-image @imgUploaded="updateNodePic($event,newNodeForm)" @func="getMsgFormSon" style="margin-bottom: 10px;"></cropper-image>
+        <cropper-image @imgUploaded="updateNodePic($event,newNodeForm)" @func="getMsgFormSon"
+                       style="margin-bottom: 10px;"></cropper-image>
         <div style="height: 80px">
           <div class="node-edit-color" style="margin: 22px;">
-            字体颜色 <el-color-picker v-model="newNodeForm.fontColor" show-alpha :predefine="predefineColors"/>
+            字体颜色
+            <el-color-picker v-model="newNodeForm.fontColor" show-alpha :predefine="predefineColors"/>
           </div>
           <div class="node-edit-color" style="margin: 22px;">
-            背景颜色 <el-color-picker v-model="newNodeForm.color" show-alpha :predefine="predefineColors"/>
+            背景颜色
+            <el-color-picker v-model="newNodeForm.color" show-alpha :predefine="predefineColors"/>
           </div>
           <div class="node-edit-color" style="margin: 22px;">
-            边框颜色 <el-color-picker v-model="newNodeForm.borderColor" show-alpha :predefine="predefineColors"/>
+            边框颜色
+            <el-color-picker v-model="newNodeForm.borderColor" show-alpha :predefine="predefineColors"/>
           </div>
         </div>
         <el-form-item prop="id">
@@ -113,7 +122,7 @@ export default {
   name: 'NodeEditor',
   components: { CropperImage },
   data () {
-    let checkNodeId = (rule, value, callback) => {
+    let checkNodeIdForUpdate = (rule, value, callback) => {
       if (!value) {
         return callback(new Error('节点名称不能为空'))
       } else {
@@ -121,6 +130,19 @@ export default {
           if (id === value)
             return callback(new Error(value + '已存在'))
         })
+        return callback()
+      }
+    }
+    let checkNodeIdForAdd = (rule, value, callback) => {
+      if (!value) {
+        return callback(new Error('节点名称不能为空'))
+      } else {
+        this.idList.forEach((id) => {
+          if (id === value)
+            return callback(new Error(value + '已存在'))
+        })
+        if (value === this.currentNode.id)
+          return callback(new Error(value + '已存在'))
         return callback()
       }
     }
@@ -142,8 +164,12 @@ export default {
         'hsla(209, 100%, 56%, 0.73)',
         '#c7158577'
       ],
-      rules: {
-        id: [{ validator: checkNodeId, trigger: 'blur' }],
+      rulesForUpdate: {
+        id: [{ validator: checkNodeIdForUpdate, trigger: 'blur' }],
+        content: [{ required: true, message: '内容不能为空', trigger: 'blur' }],
+      },
+      rulesForAdd: {
+        id: [{ validator: checkNodeIdForAdd, trigger: 'blur' }],
         content: [{ required: true, message: '内容不能为空', trigger: 'blur' }],
       },
       dialogVisible: false,
@@ -201,14 +227,35 @@ export default {
       this.dialogVisible = false
     },
     deleteNode (node) {
-      axios.delete('http://localhost:8080/node/deleteNodeByNodeId/' + node.data.id)
-          .then(({ data }) => {
-            console.log(data)
-            this.$emit('nodeUpdated')
-          })
-          .catch((err) => {
-            console.log(err)
-          })
+      this.$confirm('此操作将永久删除该该结点和与之相关的关系，是否继续？', '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(() => {
+            axios.delete('http://localhost:8080/node/deleteNodeByNodeId/' + node.data.id)
+                .then(({ data }) => {
+                  console.log(data)
+                  this.$emit('nodeUpdated')
+                  this.$message({
+                    type: 'success',
+                    message: '删除成功!'
+                  })
+                })
+                .catch((err) => {
+                  console.log(err)
+                  this.$message({
+                    message: '结点删除失败',
+                    type: 'error'
+                  })
+                })
+          }
+      ).catch(() => {
+        this.$message({
+          type: 'info',
+          message: '已取消删除'
+        })
+      })
+
     },
     addNewNode () {
       this.$refs.newNodeForm.validate((valid) => {
@@ -222,6 +269,10 @@ export default {
                   this.$emit('nodeUpdated', this.newNodeForm.id)
                   this.$refs.newNodeForm.resetFields()
                   this.newNodeDialogVisible = false
+                  this.$message({
+                    message: '添加结点成功',
+                    type: 'success'
+                  })
                 } else {
                   this.$message({
                     message: '添加结点失败',
@@ -235,10 +286,10 @@ export default {
         }
       })
     },
-    updateNodeXY(){
-      this.$emit('updateNodeXY');
+    updateNodeXY () {
+      this.$emit('updateNodeXY')
     },
-    getMsgFormSon(data){
+    getMsgFormSon (data) {
       this.msgFormSon = data
       console.log(this.msgFormSon)
     }
@@ -280,7 +331,7 @@ export default {
   }
 }
 
-.node-edit-color{
+.node-edit-color {
   float: left;
   margin: 12px;
   display: flex;
